+++ src/components/__init__.py
@@ -1,3 +1,5 @@
+from .audio_artist_and_song_name_source import AudioArtistAndSongNameSource
+
 from .file_operations import FileOperations
 
 from .text_simplifier import simplify
@@ -6,3 +8,5 @@ from .object_factory import ObjectFactory
 
 from .miscellaneous import percentage
 from .miscellaneous import get_percentage_and_amount_string
+
+from .github_repository_version_check import GithubRepositoryVersionCheck